+++ crates/unftp-sbe-fs/src/lib.rs
@@ -11,7 +11,7 @@
 //!     let ftp_home = std::env::temp_dir();
 //!     let server = libunftp::Server::with_fs(ftp_home)
 //!         .greeting("Welcome to my FTP server")
-//!         .passive_ports(50000..65535)
+//!         .passive_ports(50000..=65535)
 //!         .build()
 //!         .unwrap();
 //!
@@ -40,7 +40,7 @@ use std::{
 use tokio::io::AsyncSeekExt;
 
 #[cfg(unix)]
-use std::os::unix::fs::{MetadataExt, PermissionsExt};
+use cap_std::fs::{MetadataExt, PermissionsExt};
 
 /// The Filesystem struct is an implementation of the StorageBackend trait that keeps its files
 /// inside a specific root directory on local disk.
@@ -59,6 +59,7 @@ pub struct Filesystem {
 #[derive(Debug)]
 pub struct Meta {
     inner: cap_std::fs::Metadata,
+    target: Option<PathBuf>,
 }
 
 /// Strip the "/" prefix, if any, from a path.  Suitable for preprocessing the input pathnames
@@ -79,11 +80,11 @@ impl Filesystem {
     /// Create a new Filesystem backend, with the given root. No operations can take place outside
     /// of the root. For example, when the `Filesystem` root is set to `/srv/ftp`, and a client
     /// asks for `hello.txt`, the server will send it `/srv/ftp/hello.txt`.
-    pub fn new<P: Into<PathBuf>>(root: P) -> Self {
+    pub fn new<P: Into<PathBuf>>(root: P) -> io::Result<Self> {
         let path = root.into();
         let aa = cap_std::ambient_authority();
-        let root_fd = Arc::new(cap_std::fs::Dir::open_ambient_dir(&path, aa).unwrap());
-        Filesystem { root_fd, root: path }
+        let root_fd = Arc::new(cap_std::fs::Dir::open_ambient_dir(&path, aa)?);
+        Ok(Filesystem { root_fd, root: path })
     }
 }
 
@@ -112,7 +113,19 @@ impl<User: UserDetail> StorageBackend<User> for Filesystem {
         let fs_meta = cap_fs::symlink_metadata(self.root_fd.clone(), &path)
             .await
             .map_err(|_| Error::from(ErrorKind::PermanentFileNotAvailable))?;
-        Ok(Meta { inner: fs_meta })
+        let target = if fs_meta.is_symlink() {
+            match self.root_fd.read_link_contents(path) {
+                Ok(p) => Some(p),
+                Err(_e) => {
+                    // XXX We should really log an error here.  But a logger object is not
+                    // available.
+                    None
+                }
+            }
+        } else {
+            None
+        };
+        Ok(Meta { inner: fs_meta, target })
     }
 
     #[allow(clippy::type_complexity)]
@@ -127,8 +140,21 @@ impl<User: UserDetail> StorageBackend<User> for Filesystem {
         let fis: Vec<Fileinfo<std::path::PathBuf, Self::Metadata>> = cap_fs::read_dir(self.root_fd.clone(), path)
             .and_then(|dirent| {
                 let entry_path: PathBuf = dirent.file_name().into();
-                cap_fs::symlink_metadata(self.root_fd.clone(), path.join(entry_path.clone())).map_ok(move |meta| {
-                    let metadata = Meta { inner: meta };
+                let fullpath = path.join(entry_path.clone());
+                cap_fs::symlink_metadata(self.root_fd.clone(), fullpath.clone()).map_ok(move |meta| {
+                    let target = if meta.is_symlink() {
+                        match self.root_fd.read_link_contents(&fullpath) {
+                            Ok(p) => Some(p),
+                            Err(_e) => {
+                                // XXX We should really log an error here.  But a logger object is
+                                // not available.
+                                None
+                            }
+                        }
+                    } else {
+                        None
+                    };
+                    let metadata = Meta { inner: meta, target };
                     Fileinfo { path: entry_path, metadata }
                 })
             })
@@ -285,6 +311,10 @@ impl Metadata for Meta {
                 Permissions(0o7755)
             }
         }
+    }
+
+    fn readlink(&self) -> Option<&Path> {
+        self.target.as_deref()
     }
 }