+++ crates/unftp-sbe-gcs/examples/gcs.rs
@@ -1,6 +1,8 @@
 use clap::{Arg, Command};
+use libunftp::ServerBuilder;
 use std::{error::Error, path::PathBuf};
 use tracing::Level;
+use unftp_sbe_gcs::options::AuthMethod;
 
 // To run this example with the local fake GCS (see tests/resources/gcs_test.sh) instead of Google GCS,
 // after starting fake-gcs-server, run this example with
@@ -38,7 +40,7 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
                 .value_name("SERVICE_ACCOUNT_KEY")
                 .env("LIBUNFTP_SERVICE_ACCOUNT_KEY")
                 .help("The service account key JSON file of the Google Cloud Storage bucket to be used")
-                .required(true),
+                .required(false),
         )
         .arg(
             Arg::new(FAKE_GCS_BASE_URL)
@@ -69,41 +71,32 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
         )
         .get_matches();
 
-    let service_account_key_path = matches
-        .value_of(SERVICE_ACCOUNT_KEY)
-        .ok_or("Internal error: use of an undefined command line parameter")?;
-    let bucket_name = matches
-        .value_of(BUCKET_NAME)
-        .ok_or("Internal error: use of an undefined command line parameter")?
-        .to_owned();
-    let gcs_base_url = if let Some(base_url) = matches.value_of(FAKE_GCS_BASE_URL) {
+    let service_account_key_path = matches.get_one::<String>(SERVICE_ACCOUNT_KEY);
+    let bucket_name = matches.get_one::<String>(BUCKET_NAME).unwrap().to_owned();
+    let gcs_base_url = if let Some(base_url) = matches.get_one::<String>(FAKE_GCS_BASE_URL) {
         String::from(base_url)
     } else {
         String::from("https://www.googleapis.com")
     };
 
-    let service_account_key: Vec<u8> = tokio::fs::read(service_account_key_path).await?;
-    if let Some(ftps_certs_file) = matches.value_of(FTPS_CERTS_FILE) {
-        let ftps_key_file = matches
-            .value_of(FTPS_KEY_FILE)
-            .ok_or("Internal error: use of an undefined command line parameter")?;
-        libunftp::ServerBuilder::new(Box::new(move || {
-            unftp_sbe_gcs::CloudStorage::with_api_base(&gcs_base_url, &bucket_name, PathBuf::new(), service_account_key.clone())
-        }))
-        .ftps(ftps_certs_file, ftps_key_file)
-        .build()
-        .unwrap()
-        .listen(BIND_ADDRESS)
-        .await?;
+    let service_account_key: Option<Vec<u8>> = match service_account_key_path {
+        Some(key_path) => Some(tokio::fs::read(key_path).await?),
+        None => None,
+    };
+
+    let mut builder = ServerBuilder::new(Box::new(move || match &service_account_key {
+        Some(key) => unftp_sbe_gcs::CloudStorage::with_api_base(&gcs_base_url, &bucket_name, PathBuf::new(), key.clone()),
+        None => unftp_sbe_gcs::CloudStorage::with_api_base(&gcs_base_url, &bucket_name, PathBuf::new(), AuthMethod::None),
+    }));
+
+    builder = if let Some(ftps_certs_file) = matches.get_one::<String>(FTPS_CERTS_FILE) {
+        let ftps_key_file = matches.get_one::<String>(FTPS_KEY_FILE).unwrap();
+        builder.ftps(ftps_certs_file, ftps_key_file)
     } else {
-        libunftp::ServerBuilder::new(Box::new(move || {
-            unftp_sbe_gcs::CloudStorage::with_api_base(&gcs_base_url, &bucket_name, PathBuf::new(), service_account_key.clone())
-        }))
-        .build()
-        .unwrap()
-        .listen(BIND_ADDRESS)
-        .await?;
-    }
+        builder
+    };
+
+    builder.build().unwrap().listen(BIND_ADDRESS).await?;
 
     Ok(())
 }