+++ src/storage/error.rs
@@ -50,55 +50,55 @@ pub enum ErrorKind {
     /// Error that will cause an FTP reply code of 450 to be returned to the FTP client.
     /// The storage back-end implementation should return this if a error occurred that my be
     /// retried for example in the case where a file is busy.
-    #[display(fmt = "450 Transient file not available")]
+    #[display("450 Transient file not available")]
     TransientFileNotAvailable,
     /// Error that will cause an FTP reply code of 550 to be returned to the FTP client.
     /// The storage back-end implementation should return this if a error occurred where it doesn't
     /// make sense for it to be retried. For example in the case where a file is busy.
-    #[display(fmt = "550 Permanent file not available")]
+    #[display("550 Permanent file not available")]
     PermanentFileNotAvailable,
     /// Error that will cause an FTP reply code of 550 to be returned to the FTP client.
     /// The storage back-end implementation should return this if a error occurred where it doesn't
     /// make sense for it to be retried. For example in the case where the directory doesn't exist
-    #[display(fmt = "550 Permanent directory not available")]
+    #[display("550 Permanent directory not available")]
     PermanentDirectoryNotAvailable,
     /// Error that will cause an FTP reply code of 550 to be returned to the FTP client.
     /// The storage back-end implementation should return this if a error occurred where it doesn't
     /// make sense for it to be retried. For example in the case where file access is denied.
-    #[display(fmt = "550 The directory is not empty")]
+    #[display("550 The directory is not empty")]
     PermanentDirectoryNotEmpty,
     /// Error that will cause an FTP reply code of 550 to be returned to the FTP client.
     /// The storage back-end implementation should return this if a error occurred where it doesn't
     /// make sense for it to be retried. For example in the case where file access is denied.
-    #[display(fmt = "550 Permission denied")]
+    #[display("550 Permission denied")]
     PermissionDenied,
     /// Error that will cause an FTP reply code of 426 to be returned to the FTP client. It means the transfer was
     /// aborted, possibly by the client or because of a network issue
-    #[display(fmt = "426 Connection closed transfer aborted")]
+    #[display("426 Connection closed transfer aborted")]
     ConnectionClosed,
     /// Error that will cause an FTP reply code of 451 to be returned to the FTP client. It means
     /// the requested action was aborted due to a local error (internal storage back-end error) in
     /// processing.
-    #[display(fmt = "451 Local error")]
+    #[display("451 Local error")]
     LocalError,
     /// 551 Requested action aborted. Page type unknown.
-    #[display(fmt = "551 Page type unknown")]
+    #[display("551 Page type unknown")]
     PageTypeUnknown,
     /// 452 Requested action not taken. Insufficient storage space in system.
-    #[display(fmt = "452 Insufficient storage space error")]
+    #[display("452 Insufficient storage space error")]
     InsufficientStorageSpaceError,
     /// 552 Requested file action aborted. Exceeded storage allocation (for current directory or
     /// dataset).
-    #[display(fmt = "552 Exceeded storage allocation error")]
+    #[display("552 Exceeded storage allocation error")]
     ExceededStorageAllocationError,
     /// Error that will cause an FTP reply code of 553 to be returned to the FTP client. Its means
     /// the requested action was not taken due to an illegal file name.
-    #[display(fmt = "553 File name not allowed error")]
+    #[display("553 File name not allowed error")]
     FileNameNotAllowedError,
     /// Error that will cause an FTP reply code of 502. The indicates to the client that the command
     /// is not implemented for the storage back-end. For instance the GCS back-end don't implement
     /// RMD (remove directory) but returns this error instead from its StorageBackend::rmd
     /// implementation.
-    #[display(fmt = "502 Command not implemented")]
+    #[display("502 Command not implemented")]
     CommandNotImplemented,
 }