+++ CHANGELOG.md
@@ -1,5 +1,40 @@
 # Changelog
 
+### Upcoming release
+
+- Upgrade dependencies
+- Compiling against Rust 1.85.0
+
+### libunftp 0.20.3
+
+_tag: libunftp-0.20.3_
+
+- [#528](https://github.com/bolcom/libunftp/pull/528) Fix to enable socket reuse that caused 425 errors in passive mode
+
+## 2024-12-15 Release of all crates
+
+### libunftp 0.20.2
+
+_tag: libunftp-0.20.2_
+
+- Upgraded dependencies
+
+### unftp-auth-jsonfile v0.3.5, unftp-auth-pam v0.2.6
+
+- Compiled against libunftp v0.20.2
+- Upgraded dependencies
+
+### unftp-auth-rest v0.2.7
+
+- [520](https://github.com/bolcom/libunftp/pull/520) Allow https for auth rest url
+- Compiled against libunftp v0.20.2
+- Upgraded dependencies
+
+### unftp-sbe-fs v0.2.6, unftp-sbe-gcs v0.2.7
+
+- Compiled against libunftp v0.20.2
+- Upgraded dependencies
+
 ### libunftp 0.20.1
 
 - Fixed a build issue on Windows